Put files of the api package together
