--- internal/interactor/branch.go
@@ -1,1 +1,0 @@
-package interactor
