--- internal/interactor/chat_callback.go
@@ -1,1 +1,0 @@
-package interactor
