--- internal/interactor/chat_user.go
@@ -1,1 +1,0 @@
-package interactor
